pyproject.toml
[project]
name = "zzsn-projekt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "datasets==4.8.5",
    "diffusers>=0.37.1",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "loguru>=0.7.3",
    "torch>=2.5.1",
    "transformers>=5.7.0",
]

[tool.uv.sources]
torch = [
  { index="pytorch-cu121"},
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
